プロジェクト構成のリファクタリングとコードのモジュール化
- ディレクトリ構成を以下のように変更:
  - App: アプリのエントリーポイント
  - Views: 各画面を構成するSwiftUIのビュー
  - Models: データモデル
  - Services: サービスクラス (API通信、データ保存)
- `ContentView.swift` からデータ保存ロジックを分離し、`DataService.swift` に移動
- `APIService.swift` を追加し、API通信ロジックを移動
- コードの可読性とメンテナンス性を向上
