+++ Yumemi-intern-coding-assignment/Models/FortuneResponse.swift
