+++ Yumemi-intern-coding-assignment/Models/FortuneResponseHistory.swift
