+++ Yumemi-intern-coding-assignment/Views/DetailView.swift
