+++ Yumemi-intern-coding-assignment/Views/FortuneView.swift
